# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Job Application Tracker API",
    "version": "2.0.0",
    "description": "Job applications and interviews belonging to the signed-in GitHub user. First GitHub sign-in creates a user account. GitHub OAuth uses PKCE, and sessions use a signed HttpOnly cookie. Logout invalidates prior sessions. Open /auth/github in this browser, then return to Swagger UI to test protected routes. Existing Week 03 documents without an owner are not shown to signed-in users."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/auth/github": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Start GitHub login or account creation",
        "description": "Open this URL in a browser. GitHub asks the user to authorize the app. The first successful login creates a user document. Do not use Swagger Execute for this browser redirect.",
        "security": [],
        "responses": {
          "302": {
            "description": "Redirect to GitHub"
          },
          "503": {
            "description": "OAuth configuration is missing"
          }
        }
      }
    },
    "/auth/github/callback": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "GitHub OAuth callback",
        "description": "GitHub redirects here with a code and state. The API exchanges the code, creates or updates the user, sets an HttpOnly session cookie, and redirects to Swagger UI.",
        "security": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Signed in; redirect to API documentation"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "502": {
            "description": "GitHub login failed"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "View the signed-in user",
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log out",
        "description": "Invalidates the signed-in account sessions and clears the browser cookie.",
        "responses": {
          "204": {
            "description": "Logged out; no response body"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    }
  },
  "components": {
...
    "responses": {
      "ValidationError": {
        "description": "Invalid ID, query, JSON, or body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Validation failed",
              "details": [
                {
                  "field": "company",
                  "message": "Too small: expected string to have >=1 characters"
                }
              ]
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Sign in with GitHub first",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Sign in with GitHub first"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "login",
          "displayName",
          "avatarUrl"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "login": {
            "type": "string",
            "example": "octocat"
          },
          "displayName": {
            "type": "string",
            "example": "The Octocat"
          },
          "avatarUrl": {
            "type": "string",
            "description": "GitHub avatar URL, or an empty string if unavailable"
          }
        }
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jobapp_session",
        "description": "Set automatically after GitHub OAuth login at /auth/github. Swagger UI sends it on same-origin requests."
      }
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ]
}